+++ PrintLogUploader.py
@@ -104,6 +104,11 @@ class PrintLogUploader(QObject, Extension):
             True
         )
 
+        self._application.getPreferences().addPreference(
+            "3d_print_log/bypass_prompt",
+            False
+        )
+
         self._application.engineCreatedSignal.connect(self._onEngineCreated)
 
     def _onSendMenuButtonClicked(self):
@@ -203,11 +208,20 @@ class PrintLogUploader(QObject, Extension):
         if not hasSliced:
             return False
 
+        preferences = self._application.getInstance().getPreferences()
+        bypass_prompt = preferences.getValue(
+                "3d_print_log/bypass_prompt")
+        if (bypass_prompt):
+                return True
+
         dialog = self._createConfirmationDialog()
 
         returnValue = dialog.exec()
-
-        return returnValue == QMessageBox.Ok
+        
+        if USE_QT5:
+            return returnValue == QMessageBox.Ok
+        else:
+            return returnValue == QMessageBox.StandardButton.Ok
 
     def _hasSlicedModel(self) -> bool:
         '''Checks to see if the model has been sliced'''
@@ -224,11 +238,11 @@ class PrintLogUploader(QObject, Extension):
     def _createConfirmationDialog(self):
         '''Create a message box prompting the user if they want to send this print information.'''
         msgBox = QMessageBox()
-        msgBox.setIcon(QMessageBox.Information)
+        msgBox.setIcon(QMessageBox.Information if USE_QT5 else QMessageBox.Icon.Information)
         msgBox.setText("Would you like to send to 3Dprintlog.com?")
         msgBox.setWindowTitle("Send to 3D Print Log?")
-        msgBox.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
-        msgBox.setDefaultButton(QMessageBox.Ok)
+        msgBox.setStandardButtons((QMessageBox.Ok if USE_QT5 else QMessageBox.StandardButton.Ok) | (QMessageBox.Cancel if USE_QT5 else QMessageBox.StandardButton.Cancel))
+        msgBox.setDefaultButton(QMessageBox.Ok if USE_QT5 else QMessageBox.StandardButton.Ok)
 
         self._add3DPrintLogLogo(msgBox)
 
@@ -237,11 +251,11 @@ class PrintLogUploader(QObject, Extension):
     def _createDialog(self, text, title):
         '''Create a messsage box with a title and text'''
         msgBox = QMessageBox()
-        msgBox.setIcon(QMessageBox.Information)
+        msgBox.setIcon(QMessageBox.Information if USE_QT5 else QMessageBox.Icon.Information)
         msgBox.setText(text)
         msgBox.setWindowTitle(title)
-        msgBox.setStandardButtons(QMessageBox.Ok)
-        msgBox.setDefaultButton(QMessageBox.Ok)
+        msgBox.setStandardButtons((QMessageBox.Ok if USE_QT5 else QMessageBox.StandardButton.Ok))
+        msgBox.setDefaultButton((QMessageBox.Ok if USE_QT5 else QMessageBox.StandardButton.Ok))
 
         self._add3DPrintLogLogo(msgBox)